requirements.txt
aiohttp==3.7.4.post0
anyio==3.6.1
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
click==8.1.3
dependency-injector==4.39.1
discord.py==1.7.3
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
multidict==6.0.2
pydantic==1.9.1
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
yarl==1.7.2
